cross-platform build fixes + eden merge integration
Android:
- rename Eden application/module identifiers to suyu (settings.gradle, build.gradle)
- bump minSdk 24→26 (adaptive-icon requires ≥26)
- delete duplicate dev/suyu/suyu_emu package tree (superseded by org/yuzu tree)
- add missing string resources (home_games, early_access_*, search_*)
- rename ic_yuzu.png (duplicate) — kept only the xml adaptive icon

Linux (via WSL/Ubuntu, headless):
- SDL2 → SDL3 in conanfile.txt (matches CMakeLists find_package(SDL3))
- LINK_GROUP:RESCAN on OpenSSL/ZLIB in core to satisfy GNU ld single-pass
- pragma-disable -Wunused-variable around AMD vk_mem_alloc.h inclusion
- static_cast<u32>/static_cast<u8> for -Werror=conversion in arm_recomp.cpp

Windows: SUYU_ICON named PE resource → numeric id 1 (exporter's game
icon overwrites it instead of being shadowed).

+++ src/core/arm/recomp/arm_recomp.cpp
@@ -316,7 +316,7 @@ struct ArmRecomp::Impl {
         u32 max_index = 8192;
         if (d.strtab_va > d.symtab_va) {
             const u64 span = d.strtab_va - d.symtab_va;
-            max_index = std::min<u64>(span / 24, 65536);
+            max_index = static_cast<u32>(std::min<u64>(span / 24, 65536));
         }
         for (u32 i = 1; i < max_index; ++i) { // index 0 is always the null symbol
             const auto sym = ReadSymbol(d, i);
@@ -820,10 +820,10 @@ void ArmRecomp::SetContext(const Kernel::Svc::ThreadContext& ctx) {
     impl->ctx.x[30] = ctx.lr;
     impl->ctx.x[31] = ctx.sp;
     impl->ctx.pc = ctx.pc;
-    impl->ctx.n = (ctx.pstate >> 31) & 1;
-    impl->ctx.z = (ctx.pstate >> 30) & 1;
-    impl->ctx.c = (ctx.pstate >> 29) & 1;
-    impl->ctx.v = (ctx.pstate >> 28) & 1;
+    impl->ctx.n = static_cast<u8>((ctx.pstate >> 31) & 1);
+    impl->ctx.z = static_cast<u8>((ctx.pstate >> 30) & 1);
+    impl->ctx.c = static_cast<u8>((ctx.pstate >> 29) & 1);
+    impl->ctx.v = static_cast<u8>((ctx.pstate >> 28) & 1);
     for (size_t i = 0; i < 32; ++i) {
         impl->ctx.vreg[i][0] = ctx.v[i][0];
         impl->ctx.vreg[i][1] = ctx.v[i][1];
